introduced handling of common exceptions in service layer

+++ src/main/java/it/ned/bookmanager/service/transactional/AuthorTransactionalService.java
@@ -1,7 +1,10 @@
 package it.ned.bookmanager.service.transactional;
 
 import it.ned.bookmanager.model.Author;
+import it.ned.bookmanager.repository.AuthorRepository;
 import it.ned.bookmanager.service.AuthorService;
+import it.ned.bookmanager.service.exception.AuthorAlreadyInDatabaseException;
+import it.ned.bookmanager.service.exception.AuthorNotFoundException;
 import it.ned.bookmanager.transaction.TransactionManager;
 
 import java.util.List;
@@ -9,6 +12,9 @@ import java.util.List;
 public class AuthorTransactionalService implements AuthorService {
 
     private final TransactionManager transactionManager;
+
+    private static final String AUTHOR_NOT_FOUND_ERROR_MESSAGE = "Author with id %s not found in database.";
+    private static final String AUTHOR_ALREADY_IN_DB_ERROR_MESSAGE = "Author with id %s is already in database.";
 
     public AuthorTransactionalService(TransactionManager transactionManager) {
         this.transactionManager = transactionManager;
@@ -30,8 +36,14 @@ public class AuthorTransactionalService implements AuthorService {
 
     public void add(Author author) {
         transactionManager.doInTransaction(factory -> {
-            if (author != null)
-                factory.createAuthorRepository().add(author);
+            if (author != null) {
+                AuthorRepository authorRepository = factory.createAuthorRepository();
+                if (authorRepository.findById(author.getId()) != null)
+                    throw new AuthorAlreadyInDatabaseException(
+                            String.format(AUTHOR_ALREADY_IN_DB_ERROR_MESSAGE, author.getId())
+                    );
+                authorRepository.add(author);
+            }
             return null;
         });
     }
@@ -39,8 +51,13 @@ public class AuthorTransactionalService implements AuthorService {
     @Override
     public void delete(String authorId) {
         transactionManager.doInTransaction(factory -> {
+            AuthorRepository authorRepository = factory.createAuthorRepository();
+            if (authorRepository.findById(authorId) == null)
+                throw new AuthorNotFoundException(
+                        String.format(AUTHOR_NOT_FOUND_ERROR_MESSAGE, authorId)
+                );
             factory.createBookRepository().deleteAllBooksForAuthorId(authorId);
-            factory.createAuthorRepository().delete(authorId);
+            authorRepository.delete(authorId);
             return null;
         });
     }
